feat(instrumentation-react-native-navigation) renaming useTraceRef file
